requirements.txt
yandex-tracker-client
python-dotenv
rich
mcp>=1.0.0
